Move Footer and Header to directory `container`
