+++ src/container/Header.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
-import SearchBar from './SearchBar';
-import CitiesSuggestionBar from './CitiesSuggestionBar';
+import SearchBar from '../SearchBar';
+import CitiesSuggestionBar from '../CitiesSuggestionBar';
 
 class Header extends Component {
   render() {
